Enable local and NAS storage configuration
Allows the HomelabOS user to configure Network Attached Storage using either NFS or Samba shares. Users can define where the NAS shares are mounted on the HomelabOS host.

When a NAS is not configured, the same folder structure for the shares is used, but the folders are local to the HomelabOS install.