Fix `make terraform`
Updated `Dockerfile`. Moved jq from build-dependencies to BUILD_PACKAGES... build-dependencies are removed, but jq is needed when running `terraform make`

Updated terraform.sh to include the vault.yml file when running the terraform playbook for ansible. Also corrected the commands for terraform so the tfstate file properly resides in the settings directory.