remove task to backup database on update
this previously would try to make a copy of the database every time the
service is updated, which is not only unnecessary but would often fail
due to the directory already existing.

the user should be the one to make a backup before updating the gitea
image, updates to the service (i.e. docker-compose file) should not
touch the database or filesystem.